OAuth library added and working without use of window. Still transforms response incorrectly.
