+++ src/config.js
@@ -1,5 +1,5 @@
 export default {
-  tessellateRoot: 'http://tessellate.elasticbeanstalk.com',
+  tessellateRoot: 'http://localhost:4000',
   firebaseRoot: process.env.NODE_ENV !== 'test' ? 'https://kyper-tech.firebaseio.com/tessellate' : 'https://devshare-test.firebaseio.com/tessellate',
   oauthioKey: 'sxwuB9Gci8-4pBH7xjD0V_jooNU'
 }
