+++ src/projects/index.js
@@ -4,7 +4,7 @@ import cruder, { search } from '../utils/cruder'
 export default (username) => {
   console.log('user:', username)
   const url = username
-    ? `${config.tessellateRoot}/users/${username}/projects`
+    ? `${config.tessellateRoot}/projects/${username}`
     : `${config.tessellateRoot}/projects`
 
   const methods = {
